+++ passes/AT010/testdata/src/a/resource/resource.go
@@ -1,0 +1,5 @@
+package resource
+
+type TestCase struct {
+	IDRefreshName string
+}